mbl: Manage container uid/gid
The mbl build container shares a volume between the host environment
and the container.  The uid/gid used within the container needs to be
managed in order to avoid permission issues within both the build
itself and subsequently from the host environment when accessing the
shared volume.  The container also arranged to plumb the ssh-agent
control socket from the host environment into the container, this also
requires coordinated uid/gid.

This approach arranges to probe the current host environment uid/gid
at the point the container is launched passing the required values to
a dedicated entry point script. The entrypoint script creates the
necessary user and group on the fly before uexecing the actual
container workload with the required uid/gid using guso.

Since the actual uid is not know until run time, the existing git user
configury kludge is relocated from the Dockerfile to the entry point
script such that the necessary git configuration and github keys can
be setup for the uid.