gosu the required uid:gid rather than username
ssh-keyscan will get confused if we have multiple accounts with the
same uid (presumably it ends up looking in the wrong account 50% of
the time)

Remove the useradd -o option that allows multiple usernames to be
created for the same uid.  This ensures that if the container image
already has a uid matching the invoking users uid, we don't create
multiple accounts with the same uid.  We already have logic to deal
with the useradd failure in the case of such duplication.

Adjust the entrypoint scripts to gosu the required uid rather than
username, this necessary in light of the above change, because the
account we want might already exist with a different username.