build-mbl: Ensure we init the correct manifest before sync.
The introduction of pinning support breaks the behaviour of sync.
Pinning performs the following steps:
- repo sync
- repo manifest -r -o pinned.xml
- repo init -m pinned.xml
- repo sync

This leaves the manifest set to pinned.xml, hence on the next
invocation of build.sh the initial repo sync occurs on the old
pinned.xml rather than the unpinned manifest.

Ensure we repo init to the correct manifest before the initial repo
sync.

Signed-off-by: Marcus Shawcroft <marcus.shawcroft@arm.com>