csf-templates: Add template CSF files for signing
This patch adds CSF files for a verified boot scheme based on one of the
SRKs - minus the 'HAB' block directive - which needs to be calculated.

It's not possible with HAB v4 to use more than one SRK on multiple images
without resetting the board - so we use and reuse the same SRK here
multiple times.

The CSF is the thing that describes to the code-signing tool how to sign a
binary.

An accompanying script needs to do all sorts of extra work to populate the
HAB block correctly at build time.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>