build-mbl: Remove superflous ,stage remove in clean.
The explicit removal of ,stage is not necessary in the clean
stage. During the clean, the stage will indicate "clean", on
completion it will switch to "start".  If the clean stage is
interrupted it will be restarted from the current recorded state of
clean, this is necessary and expected behaviour.

Signed-off-by: Marcus Shawcroft <marcus.shawcroft@arm.com>