build-mbl: Support --inject-mcc
This is a very ugly short term work around.  We need to be able to
inject key material into a build.  Longer term this key material will
be injected at runtime but in the meantime this patch provides a
mechanism to pass key material encapsulated in .c files directly to
the build of cloud client.

Once we have runtime injection of key material in place this can be reverted.

Signed-off-by: Marcus Shawcroft <marcus.shawcroft@arm.com>