build-mbl: Do not 'clean' during 'start'
We have a 'clean' stage and a 'start' stage but the cleaning behaviour
is currently not well partitioned. The 'clean' stage wipes everything
then calls 'start'. The 'start' stage wipes the pinning mbl-manifest
then clones a fresh copy. The behaviour we want is that all cleaning
happens under 'clean' and 'start' simply ensures the necessary clones
are in place. This patch drops the cleaning behaviour from 'start'.

Signed-off-by: Marcus Shawcroft <marcus.shawcroft@arm.com>