imx7-code-signing: Makefile: Add initial Makefile
This patch adds an initial makefile. There are several targets for make
here. The most productive walk-through for this is given in Readme.md
already.

A number of environment variables must be set - and are checked for
depending on the build target in question

- CST_PATH
  Must point to the path where the CST binary can be found.

- OPTEE_ROOTFS
  Specifies which root filesystem binary image to sign rootfs3 or rootfs5

  Note running "make all" will sign both rootfs3 and rootfs5 files by
  default.

Build targets include

u-boot, u-boot-recovery, u-boot-scr, zImage, fdt, optee, clean, all

See Readme.md for a full list of targets

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>