imx7-code-signing: scripts/add-signed-images.sh: Add image population script
This script adds signed images into the disk image.

It *must* be run as root. This is because kpartx requires root access to
create loopback devices.

This is heavily dependent on the layout of the disk image, although
parameters are provided that can over-ride default behaviour.

- u-boot lives at a magic address at the start of the disk
  image and is dd'd there.
- the kernel and DTB live in the boot partition
- OP-TEE lives in the /lib/firmware dir within both rootfs partitions

The script will update both rootfs1 and rootfs2 with signed images specific
to the relevant rootfs. Right now only OPTEE is differentiated at the
rootfs level.

Over time as more signed images are move into individual rootfs this script
will need to churn to track.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>