imx7-code-signing: scripts/extract-unsigned-images.sh: Add image extractor script
This patch adds a script to extract binaries from an unsigned image.
Similar to the add-signed-images.sh script it must be run as root.

The expected flow is:

- make dirs
- scripts/extract-unsigned-images.sh mbl-console.wic.gz
- make all CST_PATH=path/to/cst/directory
- scripts/add-signed-images.sh

A copy of the unsigned image will be made and the keyword "signed-" will be
prepended to the copy to differentiate the signed output image from the
unsigned input image.

As part of the signing process - collateral data used to enabling signing
is removed from the signed image.

- u-boot.bin
- u-boot.cfg
- imximage.cfg.cfgtmp

The compressed signed image can then be uncompressed at a later time and
flashed to a closed i.MX part or an open i.MX part.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>