imx7-code-signing: scripts/mkimage_ver_check.sh: Add script to check u-boot version
We require u-boot with a recent provenance in order to generate a "DCD
Blocks:" log output.

The "DCD Blocks" output is used to inform the CST how to sign a binary
which the BootROM USB download protocol can authenticate. The USB
authentication logic makes use of an internal SRAM that the regular (from
MMC) authentication does not. "DCD Blocks" describes to the BootROM how
much of the downloaded image to load into SRAM.

This patch adds a script that places a version floor of 2017 on u-boot
mkimage. Previous patches added the ability to compile tip-of-tree u-boot
mkimage and use it so, this version check script exists to ensure we don't
blindly use an older/dysfunctional version.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>