Implement code for submitting jobs to LAVA
This is related to IOTMBL-421.
Add components in order to submit jobs to LAVA.  Here the list of files present
in this directory:

* Dockerfile: it is used to create a container with all dependencies needed
for running submit-to-lava.py and its tests
* lava-job-definitions directory: it contains job jinja2 templatesmwhich are
populated and submitted to LAVA
* README.md: documentation file
* requirements.txt: python dependencies. This file is automatically used by
the docker image creation. Otherwise it can be used by the user to install
dependencies in his/her own environment
* run-me.sh: bash script which encapsulates docker logic for running python
scripts
* submit-to-lava.py: python script used to submit jobs to LAVA.
* tests directory: unit tests for submit-to-lava.py (pytest is the
framework of choice).

Signed-off-by: Diego Russo <diego.russo@arm.com>