eval all readlink executions
This fix a behaviour that getopt has with paths containing ~ (tilde) and long
options that use "=" symbol.
If "--relativePath ~/path" is specified, bash will expand ~/path to
/home/user/path.
If "--relativePath=~/path" is specified, ~/path is treated and
it gets to readlink as string.
Unfortunately readlink doesn't do the tilde expansion so we need eval it in
order to perform the tilde expansion

Fix IOTMBL-943

Signed-off-by: Diego Russo <diego.russo@arm.com>