Create release manager script (#100)
Jira task  IOTMBL 838

This script is used in order to create a release branch based on a user-provided JSON configuration file. It modifies Google Repo manifest files in armmbed/mbl-manifest, and mbl-linked-repositories.conf in armmbed/meta-mbl. It then commits these two repositories, creates a new branch/tag according to the provided JSON file and pushes all modifications to GitHub.
In more details, it performs the next significant operations:  
1. Clone the mbl-manifest repository and checks out a new revision.
2. Modify specified XML manifest files present in the new branch of mbl-manifest by setting the revision for each project
3. Commit the changes made to the created mbl-manifest branch/tag and push the branch to a remote repository.
4. Does the following for each project with modified revisions in mbl-manifest Google Repo manifest files:
    *   If the repository is an Arm-MRR, clone from the revision stated in the manifest file, create a new branch/tag, and push the branch to the remote repository.
    *   If the repository is not an Arm-MRR, simply set the revision of the projects. The revision in the user configuration file must be an existing revision in the remote repository.
5. External Arm-managed repositories can also be provided in the user configuration file. After cloning the repository from a specified revision, a new branch/tag is created and pushed to the remote repository.
6. One of the configuration files in the newly created `meta-mbl` tag/branch (`mbl-linked-repositories.conf`) is optionally updated to point to a specified commit SHA in the `mbl-core` repository.

The README.md file for this script is done in Jira task IOTMBL-1040. Refer to the README.md for more details.