build-mbl: Remove Docker mounted paths (#134)
When building with Docker and using /work as a mounting point for the
--builddir we end up with broken symlinks outside of the Docker
environment. This patch removes all /work entries and replaces with
$builddir.
It also removes the /artifacts and /downloads mount points and uses
the "real" ones.

This patch also removes the deprecated --workdir option.

Part of IOTMBL-1638: Improve developer incremental build process.