build-mbl: Add tmux package to Docker image and set TERM env for interactive mode
In the current Docker image environment there is not seeting for the
DISPLAY and access to the X11 server in the host.

To overcome this restriction in the developer interactive shell mode,
bitbake can make usage of tmux with the proper TERM env set to screen.

This change was tested with following commands inside the interactive
shell:

- bitbake-layers show-appends linux-fslc
- bitbake virtual/kernel -c menuconfig
- bitbake virtual/kernel
- bitbake mbl-image-development
- repo sync

Fixes: IOTMBL-1797: mbl-tools: "bitbake virtual/kernel -c menuconfig"
fails in interactive mode