build-mbl: No longer require a full build before interactive mode (#173)
For IOTMBL-1832 Improve mbl-tools interactive mode

The script now checks for the existence of the directory containing
the BitBake layers before entering interactive mode. The script STAGE
is changed to `start` if the directory is not found. At the
`sync-pinned` STAGE, a decision is made to go to interactive mode if
it was the STAGE selected by the user.