build-mbl: Add a script to produce HTML license reports.
Introduce a tool to create HTML license reports automatically
after an mbl build.

The tool parses license.manifest.json files output during the bitbake
build, and generates an HTML report of all licenses used in the image.
New and updated licenses are flagged in the HTML output, and a difference
JSON file is produced.

This patch affects the following areas:

  * license_diff_report.py
      Add a python script license_diff_report.py to parse the JSON license
      output from the bitbake build. Also add unit tests.

  * build.sh
      Add a function to call license_diff_report.py script at build time.

      Add an --artifactory-api-key parameter, so users/CI can pass in an
      Artifactory API key to fetch an old build for comparison.

  * Dockerfile
      Edit the Dockerfile to copy the license-report script into the docker
      environment pre build.