build: Find the last modified license manifest dir when creating buid artifacts
In a function named artifact_image_manifests in build.sh, the code
checks if more than one license directory is found in the workspace
and throws an error if this is the case.

In development environments this is an issue, as bitbake creates a
new license directory for each build with a datetime in the directory
name.
So when the script reaches the artifact stage, it often finds more than
one license directory, and returns an error.

This patch modifies artifact_image_manifests to check for the last
modified license manifest directory, rather than checking only one
directory exists.