Automate board recovery.
Moved the lxc deploy and boot into a separate yaml file and include it
ins the board deploy and boot yaml files. This is because the creation
point is different for the deploy types.

Updated the BSP test pipeline to no longer need to be overlayed into the
DUT.

Created new template for recovery-deploy-boot

Created the following temporary changes:

- reated temporary templates for imx6ul and imx7d recovery device types.
- Added temporary changes into submitToLava to support the recovery board
types.

Both of these changes should be removed once all of the imx6ul and imx7d
boards are enabled for recovery mode and the "recovery" device types are
removed.