build-update-payloads.py: deal with new payload format
We're switching to a new update payload format. Payloads will no
longer be tar files so change the filename extension for the payloads we
generate.

We have tests that generate payload files manually rather than using the
create-update-payload script in meta-mbl. Rather than updating those
tests to manually create payloads using the new format, we're going to
change them to use payloads generated by this script.

We have tests that need rootfs update payloads tests that need app
update payloads. We can generate a rootfs update payload easily enough
from this script, but we can't easily generate app update payloads from
this script yet because we haven't built the apps yet when this script
runs. Get this script to generate a rootfs payload, but leave the app
payload for later when we've moved the app build to before this script
runs.

For IOTMBL-2540: Robust update: switch to using swupdate's payload
format