Automate board recovery. (#330)
* Automate board recovery.

Moved the lxc deploy and boot into a separate yaml file and include it
ins the board deploy and boot yaml files. This is because the creation
point is different for the deploy types.

Created new template for bootbomb-recovery.

Updated imx7d-pico and imx6ul-pico to use the recovery mode.