Backport lava pipeline changes to 0.9 (#336)
* Mbed crypto test plan and templates (#328)

* Mbed crypto test plans for MBL and poky
* Test job templates for pytest and poky

* Change BSP tests to use pytest. (#335)

Simplified the test pipeline to use the pytest-job template.

* Automate board recovery. (#330)

* Automate board recovery.

Moved the lxc deploy and boot into a separate yaml file and include it
ins the board deploy and boot yaml files. This is because the creation
point is different for the deploy types.

Created new template for bootbomb-recovery.

Updated imx7d-pico and imx6ul-pico to use the recovery mode.