MBL build: set up SSH and Git whenever we use a build container
We have two build scripts that use an MBL build container: build.sh and
build-update-payloads.sh. We only run ssh-setup.sh and git-setup.sh in
build.sh, but we should be setting up the container consistently for
both. If we don't set up SSH in build-update-payloads.py then we get
errors like the following when re-building mbl-cloud-client:

[2019-11-07T05:31:15.835Z] ERROR: When reparsing
/work/workspace/mbl-warrior-dev/machine-raspberrypi3-mbl/mbl-manifest/build-mbl-development/conf/../../layers/meta-mbl/meta-mbl-apps/recipes-connectivity/mbl-cloud-client/mbl-cloud-client.bb.do_populate_sysroot,
the basehash value changed from
0581bbd2389734f315aaf0cc441d5294b2214b379c8848f966f23087a696e6f6 to
6dab9a17be8427f89021bb8e12fde47eac149c38688530de4cb0717d881d6982. The
metadata is not deterministic and this needs to be fixed.

This is because tasks of mbl-cloud-client depend on SSH related
environment variables and those variables have different values if
ssh-setup.sh hasn't been run.

Run ssh-setup.sh and git-setup.sh in entrypoint.sh so that we always run
the scripts when using an MBL build container.

For IOTMBL-2619: Bitbake errors when re-building to create update
payloads.