MBL build: set up SSH and Git whenever we use a build container (#359)
We have two build scripts that use an MBL build container: build.sh and
build-update-payloads.sh. We only run ssh-setup.sh and git-setup.sh in
build.sh, but we should be setting up the container consistently for
both. 

Add a library for container setup that all build scripts can use, and add a
"--repo-host" option to build-update-payloads.py so that, like build.sh,
it can support builds that need SSH access to non-github repos.

For IOTMBL-2619: Bitbake errors when re-building to create update
payloads.