build-update-payloads: Build a multi component payload.
Tell create-update-payload to build a multi component payload for
testing.

While I'm here, tell bitbake to clean the sstate cache for the image.
This is to ensure the /etc/build file is recreated, as our rootfs
update test will check the mtime of this file and fail if it hasn't
changed since before the update.

Cleaning the cache for mbl-image-development fixes IOTMBL-2633
Creating a multi-component payload is part of IOTMBL-2292