Add option to poll for the result from lava (#341)
Sometimes setting up a callback domain:port is not feasible. However, getting the test results should be possible for the user. submit-to-lava.py polling result from lava server allows the user to get the information whether tests pass or not without callback.

This feature is disabled by default, as it lengthens the duration required to run the program quite significantly.