ci: Add clang-tidy checker.
We need a way to run static analysis on our C++ code.

This patch adds a set of shell scripts to perform clang-tidy checks on
CMake projects in a docker container.

Included:

* clang-tidy-check.sh: shell script to run clang-tidy from inside the
  container
* Dockerfile: the container specification, installs project dependencies
  for the mbl-core/updated project.
* run-me.sh: the entry point, runs the dockerr container and executes
  clang-tidy-check.sh

NOTE: to run clang-tidy checks on other CMake projects, any library
dependencies need to be added to the ci/clang-tidy-check/Dockerfile.