ci:clang-tidy-check: Add UpdateD dependencies.
UpdateD requires full C++17 support. This patch updates the build
toolchain in clang-tidy-check's docker container to support C++17.

UpdateD also depends on gRPC. We need to build gRPC from source to get
the C++ runtime. Add the fetch and build steps for gRPC and the protoc
compiler to the Dockerfile.

Out of tree builds were causing some issues with gRPC. So now we do an
in tree build, but we copy the CMake project to a sandboxed directory inside
the container before building it.