ci:clang-tidy-check: Add UpdateD dependencies. (#405)
This patch moves the dependency fetch and build steps to 
mbl-base-docker-image/Dockerfile: we have pushed a test image from this
Dockerfile to Dockerhub, in an attempt to decrease the build time of the checks.

UpdateD requires full C++17 support. This patch updates the build
toolchain in the docker container to support C++17.

UpdateD also depends on gRPC. We need to build gRPC and the
protoc compiler from source to get the C++ runtime.

Out of tree builds were causing some issues with gRPC. So now we do an
in tree build, but we copy the CMake project to a sandboxed directory inside
the container before building it.